requirements.txt
fastapi>=0.109.0
uvicorn>=0.25.0
google-generativeai>=0.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
faiss-cpu>=1.7.0
